rag-prototype/backend-automation/requirements.txt
functions-framework==3.5.0
cloudevents==1.10.1
google-cloud-pubsub==2.19.0
google-cloud-aiplatform>=1.60.0
google-cloud-storage==2.14.0
